requirements.txt
altair==6.0.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
decorator==5.2.1
executing==2.2.1
gitdb==4.0.12
gitpython==3.1.46
graphviz==0.21
idna==3.11
ipython==9.11.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.18.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pandas==2.3.3
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
prompt-toolkit==3.0.52
protobuf==6.33.5
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==23.0.1
pydeck==0.9.1
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
pyvis==0.3.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
stack-data==0.6.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
yfiles-graphs-for-streamlit==1.3.0
